2_faceguardapp/backend/requirements.txt
Flask==1.1.2
Werkzeug==1.0.1
google-cloud-storage==1.34.0
flask-cors==3.0.9
facenet-pytorch==2.5.1
torchvision==0.7.0
tensorflow==2.3.1
numpy==1.18.5
scikit-learn==0.23.2
keras==2.4.3
Flask-PyMongo==2.3.0
PyMongo==3.11.2
dnspython==2.1.0
bcrypt==3.2.0
annoy==1.16.3
torch==1.6.0